[AU-4] Requested and catch a progress for task. Example for call task and get a progress for it: http://127.0.0.1:12001/products-today http://127.0.0.1:12001/products-today/request-progress/e60011fe-abb6-4d02-b55c-0c7d0f18673d http://127.0.0.1:12001/products-today/response-progress/e60011fe-abb6-4d02-b55c-0c7d0f18673d
Reorganized a memory for aide.

Introduced an inner memory.

Added ShelveMemoBroker.

Ordered imports.

+++ aides/python/buy_on_ebay_rabbitmq/kins/share/packages/aide_server/src/aide_server/memo_brokers/filesystem.py
@@ -5,15 +5,15 @@ from sanitize_filename import sanitize
 import shutil
 from typing import Any
 
-from ..keeper_brokers.keeper_broker import KeeperBroker
 from ..log import logger
+from ..memo_brokers.memo_broker import MemoBroker
 
 
-class FilesystemKeeperBroker(KeeperBroker):
+class FilesystemMemoBroker(MemoBroker):
     def __init__(
         self,
         path_prefix: str,
-        clear: bool = True,
+        clear: bool = False,
     ):
         assert path_prefix
 
@@ -21,7 +21,7 @@ class FilesystemKeeperBroker(KeeperBroker):
 
         self.path_prefix = path_prefix
 
-        if clear:
+        if clear and os.path.exists(self.path_prefix):
             logger.info(f"🔥 Purging the storage `{self}`...")
             shutil.rmtree(self.path_prefix)
             logger.info(f"🔥 Purged the storage `{self}`.")
